refactor(auto_strategy): オペランド互換性と指標型の改善を行う
- mom指標の型をintからfloatに変更し精度を向上
- OperandGroupに価格比率スケールを追加
- 価格ベースと価格比率間の互換性を高める相関行列を追加
- フロントエンド共通コンポーネントのTableColumnsのエクスポートを見直し整理
- tsconfigにファイル名の大文字小文字の一貫性を強制する設定を追加

+++ backend/app/services/auto_strategy/core/indicator_policies.py
@@ -25,8 +25,8 @@ class Thresholds:
     ultosc_short_lt: int | None = None
     roc_long_lt: int | None = None
     roc_short_gt: int | None = None
-    mom_long_lt: int | None = None
-    mom_short_gt: int | None = None
+    mom_long_lt: float | None = None
+    mom_short_gt: float | None = None
     stoch_long_lt: int | None = None
     stoch_short_gt: int | None = None
     cmo_long_lt: int | None = None
@@ -145,7 +145,6 @@ TREND_PREF = (
     "WMA",
     "RMA",
     "HT_TRENDLINE",
-    # 新しく追加されたトレンド指標
     "FWMA",
     "SWMA",
     "VIDYA",
